+++ src/Chart/Chart.tsx
@@ -6,15 +6,24 @@
 
 import React from 'react';
 import { Bar } from 'react-chartjs-2';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { Chart } from 'chart.js';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
-import { bleChannels, Main } from 'pc-nrfconnect-shared';
+import {
+    Alert,
+    bleChannels,
+    Button,
+    getReadbackProtection,
+    Main,
+    selectedDevice,
+} from 'pc-nrfconnect-shared';
 
+import { recoverHex } from '../recoverHex';
 import {
     getAnimationDuration,
     getChannelRangeSorted,
     getLevelRangeSorted,
+    getNoDataReceived,
     getRssi,
     getRssiMax,
 } from '../reducer';
@@ -50,6 +59,10 @@ export default () => {
     const animationDuration = useSelector(getAnimationDuration);
     const channelRange = useSelector(getChannelRangeSorted);
     const [levelMin, levelMax] = useSelector(getLevelRangeSorted);
+    const device = useSelector(selectedDevice);
+    const readbackProtection = useSelector(getReadbackProtection);
+    const noData = useSelector(getNoDataReceived);
+    const dispatch = useDispatch();
 
     const convertInLevel = (v: number) => levelMin + levelMax - v;
     const limitToLevelRange = (v: number) => {
@@ -69,6 +82,17 @@ export default () => {
 
     return (
         <Main>
+            {device && noData && readbackProtection === 'protected' && (
+                <Alert variant="warning">
+                    <div className="d-flex justify-content-between">
+                        No data received for a while, maybe the firmware is
+                        wrong. Try to reflash?
+                        <Button onClick={() => dispatch(recoverHex(device))}>
+                            Recover and Program
+                        </Button>
+                    </div>
+                </Alert>
+            )}
             <div className="chart-container">
                 <Bar
                     data={{